+++ src/pearllib/package.py
@@ -80,9 +80,7 @@ def remove_packages(
         remaining_requires = set(requires).difference(args.packages)
         if remaining_requires:
             raise PackageRequiredByOtherError(
-                'Package {} cannot be removed because is required by other packages: {}'.format(
-                    package, [str(r) for r in remaining_requires]
-                )
+                f'Package {package} cannot be removed because is required by other packages: {[str(r) for r in remaining_requires]}'
             )
         remove_package(pearl_env, package, args)
 
@@ -111,11 +109,7 @@ def _run(
         reponame=package.repo_name,
     )
     cd = 'cd "$PEARL_HOME"' if cd_home else 'cd "$PEARL_PKGDIR"'
-    script = '{hookheader}\n{cd}\n{script}'.format(
-        hookheader=hookheader,
-        cd=cd,
-        script=script,
-    )
+    script = f'{hookheader}\n{cd}\n{script}'
     run_pearl_bash(script, pearl_env, input=input, enable_xtrace=enable_xtrace, enable_errexit=enable_errexit)
 
 
@@ -149,37 +143,23 @@ def info_package(pearl_env: PearlEnvironment, package: Package, args: Namespace)
     """
     requires = pearl_env.required_by(package)
     messenger.print(
-        dedent("""
-        {cyan}Name{normal}: {full_name}
-        {cyan}Description{normal}: {description}
-        {cyan}Homepage{normal}: {homepage}
-        {cyan}URL{normal}: {url}
-        {cyan}Author{normal}: {author}
-        {cyan}License{normal}: {license}
-        {cyan}Operating Systems{normal}: {os}
-        {cyan}Keywords{normal}: {keywords}
-        {cyan}Installed{normal}: {installed}
-        {cyan}Pkg Directory{normal}: {pkg_dir}
-        {cyan}Var Directory{normal}: {var_dir}
-        {cyan}Depends on{normal}: {depends}
-        {cyan}Required by{normal}: {requires}
-        """.format(
-            cyan=Color.CYAN,
-            normal=Color.NORMAL,
-            full_name=package.full_name,
-            description=package.description,
-            homepage=package.homepage,
-            url=package.url,
-            author=package.author,
-            license=package.license,
-            os=tuple(o.name.lower() for o in package.os),
-            keywords=package.keywords,
-            installed=package.is_installed(),
-            pkg_dir=package.dir,
-            var_dir=package.vardir,
-            depends=tuple(d.full_name for d in package.depends),
-            requires=tuple(r.full_name for r in requires),
-        ))
+        dedent(
+            f"""
+            {Color.CYAN}Name{Color.NORMAL}: {package.full_name}
+            {Color.CYAN}Description{Color.NORMAL}: {package.description}
+            {Color.CYAN}Homepage{Color.NORMAL}: {package.homepage}
+            {Color.CYAN}URL{Color.NORMAL}: {package.url}
+            {Color.CYAN}Author{Color.NORMAL}: {package.author}
+            {Color.CYAN}License{Color.NORMAL}: {package.license}
+            {Color.CYAN}Operating Systems{Color.NORMAL}: {tuple(o.name.lower() for o in package.os)}
+            {Color.CYAN}Keywords{Color.NORMAL}: {package.keywords}
+            {Color.CYAN}Installed{Color.NORMAL}: {package.is_installed()}
+            {Color.CYAN}Pkg Directory{Color.NORMAL}: {package.dir}
+            {Color.CYAN}Var Directory{Color.NORMAL}: {package.vardir}
+            {Color.CYAN}Depends on{Color.NORMAL}: {tuple(d.full_name for d in package.depends)}
+            {Color.CYAN}Required by{Color.NORMAL}: {tuple(r.full_name for r in requires)}
+            """
+        )
     )
     return package, requires
 
@@ -200,14 +180,10 @@ def install_package(pearl_env: PearlEnvironment, package: Package, args: Namespa
     Installs the Pearl package.
     """
     if package.is_installed():
-        raise PackageAlreadyInstalledError('{} package is already installed.'.format(package))
+        raise PackageAlreadyInstalledError(f'{package} package is already installed.')
 
     messenger.print(
-        '{cyan}* {normal}Installing {pkg} package'.format(
-            cyan=Color.CYAN,
-            normal=Color.NORMAL,
-            pkg=package,
-        )
+        f'{Color.CYAN}* {Color.NORMAL}Installing {package} package'
     )
     package.dir.mkdir(parents=True, exist_ok=True)
     if package.is_local():
@@ -215,10 +191,10 @@ def install_package(pearl_env: PearlEnvironment, package: Package, args: Namespa
     else:
         quiet = "false" if args.verbose else "true"
         script = dedent(
+            f"""
+            install_git_repo {package.url} {package.dir} "" {quiet}
             """
-            install_git_repo {pkgurl} {pkgdir} "" {quiet}
-            """
-        ).format(pkgdir=package.dir, pkgurl=package.url, quiet=quiet)
+        )
         run_pearl_bash(script, pearl_env, input=_DEFAULT_INPUT if args.no_confirm else None)
 
     package.vardir.mkdir(parents=True, exist_ok=True)
@@ -232,9 +208,9 @@ def install_package(pearl_env: PearlEnvironment, package: Package, args: Namespa
             enable_errexit=(not args.force),
         )
     except Exception as exc:
-        msg = "Error while performing {} hook function. Rolling back...".format(hook)
+        msg = f"Error while performing {hook} hook function. Rolling back..."
         if args.force:
-            message = "{}: {}".format(msg, exc.args)
+            message = f"{msg}: {exc.args}"
             if args.verbose:
                 messenger.exception(message)
             else:
@@ -253,20 +229,14 @@ def update_package(pearl_env: PearlEnvironment, package: Package, args: Namespac
     Updates the Pearl package.
     """
     if not package.is_installed():
-        raise PackageNotInstalledError('{} package has not been installed.'.format(package))
+        raise PackageNotInstalledError(f'{package} package has not been installed.')
 
     messenger.print(
-        '{cyan}* {normal}Updating {pkg} package'.format(
-            cyan=Color.CYAN,
-            normal=Color.NORMAL,
-            pkg=package,
-        )
+        f'{Color.CYAN}* {Color.NORMAL}Updating {package} package'
     )
 
     if package.has_url_changed():
-        messenger.info("The package URL for {} has changed to {}".format(
-            package.full_name, package.url
-        ))
+        messenger.info(f"The package URL for {package.full_name} has changed to {package.url}")
         messenger.info("Replacing the package with the new repository...")
         remove_package(pearl_env, package, args)
         install_package(pearl_env, package, args)
@@ -280,10 +250,10 @@ def update_package(pearl_env: PearlEnvironment, package: Package, args: Namespac
             enable_errexit=(not args.force),
         )
     except Exception as exc:
-        msg = "Error while performing {} hook function".format(hook)
+        msg = f"Error while performing {hook} hook function"
         if not args.force:
             raise HookFunctionError(msg) from exc
-        message = "{}: {}".format(msg, exc.args)
+        message = f"{msg}: {exc.args}"
         if args.verbose:
             messenger.exception(message)
         else:
@@ -294,10 +264,10 @@ def update_package(pearl_env: PearlEnvironment, package: Package, args: Namespac
     else:
         quiet = "false" if args.verbose else "true"
         script = dedent(
+            f"""
+            update_git_repo {package.dir} "" {quiet}
             """
-            update_git_repo {pkgdir} "" {quiet}
-            """
-        ).format(pkgdir=package.dir, quiet=quiet)
+        )
         run_pearl_bash(script, pearl_env, input=_DEFAULT_INPUT if args.no_confirm else None)
 
     hook = 'post_update'
@@ -309,10 +279,10 @@ def update_package(pearl_env: PearlEnvironment, package: Package, args: Namespac
             enable_errexit=(not args.force),
         )
     except Exception as exc:
-        msg = "Error while performing {} hook function".format(hook)
+        msg = f"Error while performing {hook} hook function"
         if not args.force:
             raise HookFunctionError(msg) from exc
-        message = "{}: {}".format(msg, exc.args)
+        message = f"{msg}: {exc.args}"
         if args.verbose:
             messenger.exception(message)
         else:
@@ -324,14 +294,10 @@ def remove_package(pearl_env: PearlEnvironment, package: Package, args: Namespac
     Remove the Pearl package.
     """
     if not package.is_installed():
-        raise PackageNotInstalledError('{} package has not been installed.'.format(package))
+        raise PackageNotInstalledError(f'{package} package has not been installed.')
 
     messenger.print(
-        '{cyan}* {normal}Removing {pkg} package'.format(
-            cyan=Color.CYAN,
-            normal=Color.NORMAL,
-            pkg=package,
-        )
+        f'{Color.CYAN}* {Color.NORMAL}Removing {package} package'
     )
 
     hook = 'pre_remove'
@@ -343,10 +309,10 @@ def remove_package(pearl_env: PearlEnvironment, package: Package, args: Namespac
             enable_errexit=(not args.force),
         )
     except Exception as exc:
-        msg = "Error while performing {} hook function".format(hook)
+        msg = f"Error while performing {hook} hook function"
         if not args.force:
             raise HookFunctionError(msg) from exc
-        message = "{}: {}".format(msg, exc.args)
+        message = f"{msg}: {exc.args}"
         if args.verbose:
             messenger.exception(message)
         else:
@@ -355,14 +321,11 @@ def remove_package(pearl_env: PearlEnvironment, package: Package, args: Namespac
     shutil.rmtree(str(package.dir))
 
 
-def list_packages(pearl_env: PearlEnvironment, args: Namespace):
-    """
-    Lists or searches Pearl packages.
-    """
+def _list_packages(pearl_env: PearlEnvironment, args: Namespace):
     uninstalled_packages = []
     installed_packages = []
     pattern = args.pattern if hasattr(args, 'pattern') else ".*"
-    regex = re.compile('{}'.format(pattern), flags=re.IGNORECASE)
+    regex = re.compile(f'{pattern}', flags=re.IGNORECASE)
     for _, repo_packages in pearl_env.packages.items():
         for _, package in repo_packages.items():
             if not regex.search(package.full_name) \
@@ -373,32 +336,30 @@ def list_packages(pearl_env: PearlEnvironment, args: Namespace):
                 installed_packages.append(package)
             else:
                 uninstalled_packages.append(package)
+    if args.installed_only:
+        return installed_packages
+    return uninstalled_packages + installed_packages
 
-    total_packages = uninstalled_packages + installed_packages
+
+def list_packages(pearl_env: PearlEnvironment, args: Namespace):
+    """
+    Lists or searches Pearl packages.
+    """
+    total_packages = _list_packages(pearl_env, args)
+    if args.dependency_tree:
+        total_packages = list(closure_dependency_tree(total_packages))
+
     for package in total_packages:
         if args.package_only:
-            template = "{reponame}/{package}"
             messenger.print(
-                template.format(
-                    reponame=package.repo_name,
-                    package=package.name,
-                )
+                f"{package.repo_name}/{package.name}"
             )
         else:
             label = "[installed]" if package.is_installed() else ""
-            template = "{pink}{{reponame}}/{cyan}{{package}} {{installed}}{normal}".format(
-                pink=Color.PINK,
-                cyan=Color.CYAN,
-                normal=Color.NORMAL,
-            )
             messenger.print(
-                template.format(
-                    reponame=package.repo_name,
-                    package=package.name,
-                    installed=label,
-                )
+                f"{Color.PINK}{package.repo_name}/{Color.CYAN}{package.name} {label}{Color.NORMAL}"
             )
-            messenger.print("    {}".format(package.description))
+            messenger.print(f"    {package.description}")
     return total_packages
 
 
@@ -410,11 +371,11 @@ def create_package(pearl_env: PearlEnvironment, args: Namespace):
     pearl_config_template = static / 'templates/pearl-config.template'
     dest_pearl_config = args.dest_dir / 'pearl-config'
     if dest_pearl_config.exists():
-        raise PackageCreateError('The pearl-config directory already exists in {}'.format(args.dest_dir))
+        raise PackageCreateError(f'The pearl-config directory already exists in {args.dest_dir}')
     shutil.copytree(str(pearl_config_template), str(dest_pearl_config))
 
-    messenger.info('Updating {} to add package in local repository...'.format(pearl_env.config_filename))
+    messenger.info(f'Updating {pearl_env.config_filename} to add package in local repository...')
     with pearl_env.config_filename.open('a') as pearl_conf_file:
-        pearl_conf_file.write('PEARL_PACKAGES["{}"] = {{"url": "{}"}}\n'.format(args.name, args.dest_dir))
+        pearl_conf_file.write(f'PEARL_PACKAGES["{args.name}"] = {{"url": "{args.dest_dir}"}}\n')
 
     messenger.info('Run "pearl search local" to see your new local package available')